feat(data): 6a — property header on row pages; DataRowFields extraction; rows-by-id fetch
The plan's level-2 row page, completed: cells render as a collapsible
Properties strip above the promoted page's note body. The peek's field
stack is extracted into DataRowFields — one set of editors serving the
peek, the split pane, and now the header, so a field fix lands on every
surface at once. The header fetches its row hydrated through a new
rows `?ids=` branch (loadRowsByIds — also the missing piece behind
?row= deep links beyond page 1), commits through the same PATCH the
grid uses, refetches after every commit so refs stay honest, and a
primary-cell edit dispatches content-updated so page title, tab, and
tree label follow the server-side title sync.
